Explicitly set optimizations :none
This is actually going to annoy some people and I will likely remove,
but it served as a learning moment, so it gets to stay for now.  The
idea is that I was able to get things to run, kinda, when I specified
optimizations to :simplified.  It now dawned on me that :simplified
creates one single file and puts everything in it which is why jest
would now pass the test.  This made me refocus on my mental model which
led me to explicitly set :none and get working through the problem again